Faster AutoPID
Changed millis() to micros() to allow the function to run as fast as one can get it to - e.g. on a Due. 
For PID loops, the essential performance element is the speed at which the loop can be closed. It makes little sense to limit this to 1ms. 
Millis() will cycle over after 70 minutes, however the code looks like it will only skip one PID execution at that time.